+++ packages/act_thumbhash_flutter/lib/act_thumbhash_flutter.dart
@@ -1,3 +1,1 @@
-library act_thumbhash_flutter;
-
 export 'src/thumbhash_image_provider.dart';